+++ src/page/user/costCall/MovingAddressModal.tsx
@@ -1,11 +1,11 @@
 import React, { useState } from 'react';
 import cn from 'classnames';
-import { fetchAddress } from '../../../../lib/api/kakao';
+import { fetchAddress } from '../../../lib/api/kakao';
 import styles from './MovingAddressModal.module.css';
-import icSearchLarge from '../../../../assets/icons/ic_search_large.svg';
-import icXCircleLarge from '../../../../assets/icons/ic_x_circle_large.svg';
-import icXLarge from '../../../../assets/icons/ic_x_large.svg';
-import Button from '../../../../components/btn/Button';
+import icSearchLarge from '../../../assets/icons/ic_search_large.svg';
+import icXCircleLarge from '../../../assets/icons/ic_x_circle_large.svg';
+import icXLarge from '../../../assets/icons/ic_x_large.svg';
+import Button from '../../../components/btn/Button';
 
 interface AddressValues {
   road_address: {
