+++ src/page/user/costCall/MovingType.tsx
@@ -1,12 +1,12 @@
 import { useState } from 'react';
 import cn from 'classnames';
-import pageStyles from '../index.module.css';
+import pageStyles from './index.module.css';
 import styles from './MovingType.module.css';
-import icCheckCircleLarge from '../../../../assets/icons/ic_check_circle_large.svg';
-import icCheckCircleEmptyLarge from '../../../../assets/icons/ic_check_circle_empty_medium.svg';
+import icCheckCircleLarge from '../../../assets/icons/ic_check_circle_large.svg';
+import icCheckCircleEmptyLarge from '../../../assets/icons/ic_check_circle_empty_medium.svg';
 // import icCheckCircleMedium from '../../../../assets/icons/ic_check_circle_medium.svg';
 // import icCheckCircleEmptyMedium from '../../../../assets/icons/ic_check_circle_empty_large.svg';
-import Button from '../../../../components/btn/Button';
+import Button from '../../../components/btn/Button';
 
 type TypeValue = string | null;
